// src/models/products.model.ts
export interface PRODUCTS  {
    products_id?:  number;
    products_quantity:  number;
    products_model:  string;
    products_image:  string;
    products_price:  number;
    products_date_added:  string;
    products_last_modified:  string;
    products_date_available:  string;
    products_weight:  string;
    products_status:  number;
    products_tax_class_id:  number;
    manufacturers_id:  number;
    products_ordered:  number
}